Working on delete button refresh issue. Half fix of reloading the page; half fix of v-for key on note class not just idx

+++ app/feedback/src/components/snote.ts
@@ -31,10 +31,29 @@ export default class Snote extends Vue {
   //   this.note.deleted = del;
   // }
 
-  public delete_snote( e: any ){ //
+  public async delete_snote( ){ 
     log.info('Calling delete_snote from Snote component');
-    this.$store.dispatch( 'feedback/delete_snote', this.note_idx )
-    //return this.$store.state.feedback.delete_snote; //idx
+    const confirm_delete = confirm("Are you sure you want to delete the note:\n\""+this.get_note.content+"\"");
+    if( confirm_delete ){
+      this.$store.dispatch( 'feedback/delete_snote', this.note_idx )
+      // force the component to rerender, also didn't work
+      // this.$forceUpdate();
+
+      // splice the value out of the array 
+      // this was tried morning of 5/31 and it didn't solve the refresh issue. 
+      const index = this.$store.state.feedback.snotes.indexOf( this.get_note );
+      if( index >= 0 ){
+        const bef_arr = Array.from(this.$store.state.feedback.snotes);
+        this.$store.state.feedback.snotes.splice( index, 1 );
+        const after_arr = Array.from(this.$store.state.feedback.snotes);
+      }
+      // changing the v-for key (in assignment) from curr_note.idx to curr_note.get_note now updates when you leave and come back to the page
+      // fails to do it instantly still.
+      
+      // refresh the page
+      location.reload();  
+    }
+
   }
 
 }
